+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added `url` field to container image description
 * added `defaultValue` to action parameter description
+* added `includeConnected` parameter to `GET /containers` route, analogous to `GET /agents` and others
 * compatibility with Java 25